Port async_test
